requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

# Paquets PyTorch
torch
torchvision
torchaudio

# Autres librairies
jupyter
pandas
numpy
matplotlib
pillow